Change to float

+++ events.go
@@ -22,9 +22,9 @@ type Event struct {
 }
 
 type Data struct {
-	Pinnumber int  `json:"pinnumber"`
-	Active    bool `json:"active"`
-	Status    bool `json:"status"`
+	Pinnumber float64 `json:"pinnumber"`
+	Active    bool    `json:"active"`
+	Status    bool    `json:"status"`
 }
 
 func NewEventFromRaw(rawData []byte) (*Event, error) {
